backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
pytest==6.2.5
python-dotenv==0.19.0
pydantic>=1.8.0
requests>=2.25.1
aiohttp>=3.7.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 